+++ RELEASE_NOTES.md
@@ -44,6 +44,9 @@ files.
 
 * IMPROVED: Performance when compiling entity ground terms.
 
+* IMPROVED: Performance when compiling entities terms that are not
+term-expanded.
+
 * FIXED: Catch any errors during the compilation of the `begin_of_file`
 virtual term.